Small refactor to registering shared files, code tidy,
