+++ lib/directoryService.js
@@ -82,17 +82,17 @@ export async function getMasterEndpoint() {
 }
 
 /**
- * Gets all of the currently public files from all users
+ * Gets all of the public files for a certain user
  * GET <DIRECTORY_SERVICE>/publicFiles
  */
-export async function getAllPublicFiles() {
-  const { ok, status, response } = await makeRequest(`${ENDPOINTS.DIRECTORY_SERVICE}/publicFiles`, 'get');
+export async function getPublicFilesForUser(email) {
+  const { ok, status, response } = await makeRequest(`${ENDPOINTS.DIRECTORY_SERVICE}/publicFiles/${email}`, 'get');
   if(!ok) {
     logError(status, response);
   }
 
   console.log(`Public files available: `);
-  console.log(response);
+  console.log(JSON.stringify(response));
   return response;
 }
 
@@ -103,9 +103,9 @@ export async function getAllPublicFiles() {
  * body: { clientFileName, remoteNodeAddress, remoteFileId }
  * This post body can be formed after fetching all public files and deciding which file we want
  */
-export async function registerSharedPublicFile(fileNameForUs, remoteNodeAddress, remoteFileId) {
+export async function registerSharedPublicFile(fileNameForUs, ownerId, remoteFileId) {
   const {ok, status, response } = await makeRequest(`${ENDPOINTS.DIRECTORY_SERVICE}/sharedFile`, 'post',
-    {clientFileName: fileNameForUs, remoteNodeAddress, remoteFileId});
+    {clientFileName: fileNameForUs, ownerId, remoteFileId});
 
   if(!ok) {
     logError(status, response);
