+++ lib/lockingService.js
@@ -13,7 +13,7 @@ export async function acquireLock(_id) {
     logError(status, response);
   }
 
-  console.log(`Acquired Lock ${response.lock}`);
+  console.log(`Acquired Lock for ${_id}`);
   return response.lock;
 }
