+++ lib/remoteFileSystem.js
@@ -26,9 +26,8 @@ export async function updateRemoteFile(filename) {
 
   const { _id } = await getRemoteFileInfo(filename);
   const endpoint = await getMasterEndpoint();
-  console.log(`Endpoint: ${endpoint}`);
   const lock = await acquireLock(_id);
-  const response = await uploadFile(endpoint, filename, { lock });
+  const response = await uploadFile(`${endpoint}/${_id}`, filename, { lock });
   console.log(`Result of updateRemoteFile: `);
   console.log(`${response}`);
 
@@ -51,6 +50,9 @@ export async function renameRemoteFile(oldFileName, newFileName) {
   const response = await uploadFile(endpoint, newFileName, { lock });
   console.log(`Result of renameRemoteFile: `);
   console.log(`${response}`);
+
+  console.log(`Releasing lock`);
+  await releaseLock(_id, lock);
 }
