+++ lib/util.js
@@ -51,7 +51,6 @@ export async function makeRequest(endpoint, method, body) {
   let response;
   if(body) {
     if(sessionKey) {
-      console.log(`Making request to ${endpoint} - encrypting`);
       const encrypted = encrypt(JSON.stringify(body), sessionKey);
       response = await fetch(endpoint, {
         method,
