--- users/Frank/remoteFiles/frank.txt
@@ -1,1 +1,0 @@
-Updated at 12/8/2017, 2:26:04 PM
