--- users/Frank/remoteFiles/franks_stefano.txt
@@ -1,1 +1,0 @@
-Frank: Updated at 12/8/2017, 4:59:09 PM
